refactor(hickey): split useCommentMode out of useComments
useComments.ts braided two independently-volatile concerns: a global
session-wide UI toggle and per-repoRoot comment buckets. They share no
state or lifecycle. Moving the toggle to its own module unfuses the
change axes — a future per-pane toggle or backend swap touches one
file instead of two.
